Added support for resarting from pre-existing checkpoint files.
Adds a restart flag that checks the output_directory for pre-existing sire systems, if they are present they are loaded and used as a starting point for the runner. In this case runtime represents the total time for the system. Also adds a supress_overwrite_warning flag that supresses the need for the user to press enter before files are overwritten in the case of restart=False.