Config now encoded in sire checkpoint files
Added extra parameter to config.as_dict to make it equivalent to sire outputs.

Restarts are checked against previous config, if this config doesn't exist the runner checks the checkpoint file for lambda=0 for the previous config.

yaml safe load throws a different exception on windows - this was causing the test to fail. Should now be fixed.