Disabled equilibration for systems starting from checkpoint files.
Config will now throw a warning if restart is True and equilibration time is greater than 0, in this case the only systems that will be equilibrated will be those starting from scratch.

Dynamics._run now takes a is_restart argument that tells it whether the system it is simulating is starting from scratch or from a restart file.